refactor: Separate QuestExtraItem from QuestItem
